+++ src/grib2/section7/section7.rs
@@ -3,7 +3,7 @@ use crate::grib2::common::grib2_error::Grib2Error;
 pub struct Section7 {
     pub length: u32,
     pub section_number: u8,
-    pub data_points: Vec<f32>
+    pub data_points: Vec<u16>
 }
 
 
@@ -13,7 +13,7 @@ impl Section7 {
     pub fn new(
         length: u32,
         section_number: u8,
-        data_points: Vec<f32>
+        data_points: Vec<u16>
     ) -> Result<Section7, Grib2Error> {
         if section_number != SECTION_NUMBER {
             return Err(Grib2Error::InvalidData(
